+++ src/main/java/com/project/doongdoong/domain/analysis/domain/AnalysisEntity.java
@@ -2,7 +2,7 @@ package com.project.doongdoong.domain.analysis.domain;
 
 import com.project.doongdoong.domain.answer.domain.AnswerEntity;
 import com.project.doongdoong.domain.counsel.model.Counsel;
-import com.project.doongdoong.domain.question.model.Question;
+import com.project.doongdoong.domain.question.domain.Question;
 import com.project.doongdoong.domain.user.model.User;
 import com.project.doongdoong.global.common.BaseEntity;
 import jakarta.persistence.*;
