+++ src/main/java/com/project/doongdoong/domain/question/application/port/in/QuestionProvidable.java
@@ -1,6 +1,6 @@
-package com.project.doongdoong.domain.question.service;
+package com.project.doongdoong.domain.question.application.port.in;
 
-import com.project.doongdoong.domain.question.model.Question;
+import com.project.doongdoong.domain.question.domain.Question;
 
 import java.util.List;
