+++ src/main/java/com/project/doongdoong/domain/question/domain/Questions.java
@@ -1,4 +1,4 @@
-package com.project.doongdoong.domain.question.model;
+package com.project.doongdoong.domain.question.domain;
 
 import lombok.Getter;
