+++ src/main/java/com/project/doongdoong/domain/voice/repository/VoiceRepository.java
@@ -1,6 +1,6 @@
 package com.project.doongdoong.domain.voice.repository;
 
-import com.project.doongdoong.domain.question.model.QuestionContent;
+import com.project.doongdoong.domain.question.domain.QuestionContent;
 import com.project.doongdoong.domain.voice.model.Voice;
 import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
