+++ src/main/java/com/project/doongdoong/domain/voice/service/VoiceService.java
@@ -1,6 +1,6 @@
 package com.project.doongdoong.domain.voice.service;
 
-import com.project.doongdoong.domain.question.model.QuestionContent;
+import com.project.doongdoong.domain.question.domain.QuestionContent;
 import com.project.doongdoong.domain.voice.dto.response.VoiceDetailResponseDto;
 import com.project.doongdoong.domain.voice.model.Voice;
 import org.springframework.web.multipart.MultipartFile;
