+++ src/test/java/com/project/doongdoong/domain/user/repository/spring/UserRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.project.doongdoong.domain.user.repository.spring;
 
 import com.project.doongdoong.domain.analysis.domain.AnalysisEntity;
-import com.project.doongdoong.domain.question.model.Question;
+import com.project.doongdoong.domain.question.domain.Question;
 import com.project.doongdoong.domain.user.model.SocialType;
 import com.project.doongdoong.domain.user.model.User;
 import com.project.doongdoong.domain.user.repository.UserRepository;
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-import static com.project.doongdoong.domain.question.model.QuestionContent.*;
+import static com.project.doongdoong.domain.question.domain.QuestionContent.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
